requirements.txt
pandas
numpy
saeaborn
torch
random
warnings
logging
scikit-learn
sklearn
statsmodels
matplotlib
jupyter
flaskprophet
pmdarima
pickle


